add cloudflare testing
