+++ .github/workflows/test.yml
@@ -129,3 +129,22 @@ jobs:
               env:
                   CONNECTION_TYPE: turso
               run: npm run test:connection
+
+    test_cloudflare:
+        name: 'Cloudflare D1 Connection'
+        runs-on: ubuntu-latest
+        needs: build
+
+        steps:
+            - uses: actions/checkout@v4
+
+            - name: Install modules
+              run: npm install
+
+            - name: Run tests
+              env:
+                  CONNECTION_TYPE: cloudflare
+                  CLOUDFLARE_API_KEY: ${{ secrets.CLOUDFLARE_API_KEY }}
+                  CLOUDFLARE_ACCOUNT_ID: ${{ secrets.CLOUDFLARE_ACCOUNT_ID }}
+                  CLOUDFLARE_DATABASE_ID: ${{ secrets.CLOUDFLARE_DATABASE_ID }}
+              run: npm run test:connection
